simplify app scope tests
